+++ docs/src/components/sidebarIndex.js
@@ -9,8 +9,7 @@ export type sidebarIndexType = Array<{|
 
 const componentSubSectionPages = {
   displayDate: ['Avatar', 'AvatarPair', 'Badge', 'GroupAvatar', 'Table'],
-  feedBack: ['Modal', 'Pulsar', 'Spinner', 'Toast'],
-  foundations: ['Heading', 'Icon', 'Text'],
+  feedBack: ['Modal', 'Pulsar', 'Spinner', 'Toast', 'Tooltip'],
   forms: [
     'Button',
     'Checkbox',
@@ -29,6 +28,7 @@ const componentSubSectionPages = {
     'Tooltip',
     'Typeahead',
   ],
+  foundation: ['Heading', 'Icon', 'Text'],
   layout: [
     'Box',
     'Card',
@@ -44,6 +44,7 @@ const componentSubSectionPages = {
   ],
   media: ['Image', 'Letterbox', 'Mask', 'Video'],
   navigation: ['Link', 'SegmentedControl', 'Tabs'],
+  setup: ['GestaltProvider'],
 };
 
 // sidebarIndex is the source of truth for the sidebar documentation menu.
@@ -63,10 +64,19 @@ const sidebarIndex: sidebarIndexType = [
     sectionName: 'Getting Started',
     pages: ['Installation'],
   },
+
   {
     sectionPathname: 'components',
     sectionName: 'Components',
     subsections: [
+      {
+        subsectionName: 'Foundation',
+        pages: componentSubSectionPages.foundation,
+      },
+      {
+        subsectionName: 'Setup',
+        pages: componentSubSectionPages.setup,
+      },
       {
         subsectionName: 'Data Display',
         pages: componentSubSectionPages.displayDate,
@@ -75,10 +85,7 @@ const sidebarIndex: sidebarIndexType = [
         subsectionName: 'Feedback',
         pages: componentSubSectionPages.feedBack,
       },
-      {
-        subsectionName: 'Foundation',
-        pages: componentSubSectionPages.foundations,
-      },
+
       {
         subsectionName: 'Forms',
         pages: componentSubSectionPages.forms,