+++ packages/gestalt/src/Video.js
@@ -3,6 +3,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import VideoControls from './VideoControls.js';
+import { ColorSchemeProvider } from './contexts/ColorScheme.js';
 import styles from './Video.css';
 import Box from './Box.js';
 
@@ -11,6 +12,8 @@ type Source =
   | Array<{| type: 'video/m3u8' | 'video/mp4' | 'video/ogg', src: string |}>;
 
 type Props = {|
+  accessibilityHideCaptionsLabel?: string,
+  accessibilityShowCaptionsLabel?: string,
   accessibilityMaximizeLabel: string,
   accessibilityMinimizeLabel: string,
   accessibilityMuteLabel: string,
@@ -59,6 +62,7 @@ type State = {|
   currentTime: number,
   duration: number,
   fullscreen: boolean,
+  captionsButton: 'enabled' | 'disabled' | null,
 |};
 
 // For more information on fullscreen and vendor prefixes see
@@ -157,6 +161,8 @@ export default class Video extends React.PureComponent<Props, State> {
   player: ?HTMLDivElement;
 
   static propTypes = {
+    accessibilityHideCaptionsLabel: PropTypes.string,
+    accessibilityShowCaptionsLabel: PropTypes.string,
     accessibilityMaximizeLabel: PropTypes.string,
     accessibilityMinimizeLabel: PropTypes.string,
     accessibilityMuteLabel: PropTypes.string,
@@ -205,13 +211,14 @@ export default class Video extends React.PureComponent<Props, State> {
     playbackRate: 1,
     playing: false,
     preload: 'auto',
-    volume: 1,
+    volume: 0,
   };
 
   state: State = {
     currentTime: 0,
     duration: 0,
     fullscreen: false,
+    captionsButton: this.props.captions ? 'enabled' : null,
   };
 
   /**
@@ -219,7 +226,7 @@ export default class Video extends React.PureComponent<Props, State> {
    */
 
   componentDidMount() {
-    const { playbackRate, playing, volume } = this.props;
+    const { captions, playbackRate, playing, volume } = this.props;
     // Set up event listeners to catch backdoors in fullscreen
     // changes such as using the ESC key to exit
     if (typeof document !== 'undefined') {
@@ -234,6 +241,15 @@ export default class Video extends React.PureComponent<Props, State> {
     // Simulate an autoplay effect if the component
     if (playing) {
       this.play();
+    }
+
+    if (
+      captions &&
+      this.video &&
+      this.video.textTracks &&
+      this.video.textTracks[0]
+    ) {
+      this.video.textTracks[0].mode = 'showing';
     }
   }
 
@@ -314,7 +330,12 @@ export default class Video extends React.PureComponent<Props, State> {
   // Play the video
   play: () => void = () => {
     if (this.video) {
-      this.video.play();
+      const isPlaying =
+        this.video.currentTime > 0 &&
+        !this.video.paused &&
+        !this.video.ended &&
+        this.video.readyState > 2;
+      if (!isPlaying) this.video.play();
     }
   };
 
@@ -322,6 +343,16 @@ export default class Video extends React.PureComponent<Props, State> {
   seek: (time: number) => void = time => {
     if (this.video) {
       this.video.currentTime = time;
+    }
+  };
+
+  // Toggle captions on/off
+  toggleCaptions: () => void = () => {
+    const [videoTrack] = this.video?.textTracks || [];
+    if (videoTrack) {
+      const isShowing = videoTrack.mode === 'showing';
+      videoTrack.mode = isShowing ? 'disabled' : 'showing';
+      this.setState({ captionsButton: isShowing ? 'disabled' : 'enabled' });
     }
   };
 
@@ -480,7 +511,7 @@ export default class Video extends React.PureComponent<Props, State> {
       src,
       volume,
     } = this.props;
-    const { currentTime, duration, fullscreen } = this.state;
+    const { currentTime, duration, fullscreen, captionsButton } = this.state;
 
     const paddingBottom = (fullscreen && '0') || `${(1 / aspectRatio) * 100}%`;
 
@@ -490,57 +521,68 @@ export default class Video extends React.PureComponent<Props, State> {
         className={styles.player}
         style={{ paddingBottom, height: fullscreen ? '100%' : 0 }}
       >
-        <video
-          autoPlay={playing}
-          loop={loop}
-          muted={volume === 0}
-          playsInline={playsInline}
-          poster={poster}
-          preload={preload}
-          src={typeof src === 'string' ? src : undefined}
-          ref={this.setVideoRef}
-          className={styles.video}
-          onCanPlay={this.handleCanPlay}
-          onDurationChange={this.handleDurationChange}
-          onEnded={this.handleEnded}
-          onSeeked={this.handleSeek}
-          onTimeUpdate={this.handleTimeUpdate}
-          onProgress={this.handleProgress}
-        >
-          {Array.isArray(src) &&
-            src.map(source => (
-              <source key={source.src} src={source.src} type={source.type} />
-            ))}
-          <track kind="captions" src={captions} />
-        </video>
-        {children && (
-          <Box position="absolute" top left bottom right overflow="hidden">
-            {children}
-          </Box>
-        )}
-        {/* Need to use full path for these props so Flow can infer correct subtype */}
-        {this.props.controls && (
-          <VideoControls
-            accessibilityMaximizeLabel={this.props.accessibilityMaximizeLabel}
-            accessibilityMinimizeLabel={this.props.accessibilityMinimizeLabel}
-            accessibilityMuteLabel={this.props.accessibilityMuteLabel}
-            accessibilityPauseLabel={this.props.accessibilityPauseLabel}
-            accessibilityPlayLabel={this.props.accessibilityPlayLabel}
-            accessibilityUnmuteLabel={this.props.accessibilityUnmuteLabel}
-            currentTime={currentTime}
-            duration={duration}
-            fullscreen={fullscreen}
-            onPlay={this.handlePlay}
-            onPlayheadDown={this.handlePlayheadDown}
-            onPlayheadUp={this.handlePlayheadUp}
-            onPause={this.handlePause}
-            onFullscreenChange={this.toggleFullscreen}
-            onVolumeChange={this.handleVolumeChange}
-            playing={playing}
-            seek={this.seek}
-            volume={volume}
-          />
-        )}
+        <ColorSchemeProvider id="Video" colorScheme="light">
+          <video
+            autoPlay={playing}
+            crossOrigin="anonymous"
+            loop={loop}
+            muted={volume === 0}
+            playsInline={playsInline}
+            poster={poster}
+            preload={preload}
+            src={typeof src === 'string' ? src : undefined}
+            ref={this.setVideoRef}
+            className={styles.video}
+            onCanPlay={this.handleCanPlay}
+            onDurationChange={this.handleDurationChange}
+            onEnded={this.handleEnded}
+            onSeeked={this.handleSeek}
+            onTimeUpdate={this.handleTimeUpdate}
+            onProgress={this.handleProgress}
+          >
+            {Array.isArray(src) &&
+              src.map(source => (
+                <source key={source.src} src={source.src} type={source.type} />
+              ))}
+            <track kind="captions" src={captions} />
+          </video>
+          {children && (
+            <Box position="absolute" top left bottom right overflow="hidden">
+              {children}
+            </Box>
+          )}
+          {/* Need to use full path for these props so Flow can infer correct subtype */}
+          {this.props.controls && (
+            <VideoControls
+              accessibilityHideCaptionsLabel={
+                this.props.accessibilityHideCaptionsLabel || ''
+              }
+              accessibilityShowCaptionsLabel={
+                this.props.accessibilityShowCaptionsLabel || ''
+              }
+              accessibilityMaximizeLabel={this.props.accessibilityMaximizeLabel}
+              accessibilityMinimizeLabel={this.props.accessibilityMinimizeLabel}
+              accessibilityMuteLabel={this.props.accessibilityMuteLabel}
+              accessibilityPauseLabel={this.props.accessibilityPauseLabel}
+              accessibilityPlayLabel={this.props.accessibilityPlayLabel}
+              accessibilityUnmuteLabel={this.props.accessibilityUnmuteLabel}
+              captionsButton={captionsButton}
+              currentTime={currentTime}
+              duration={duration}
+              fullscreen={fullscreen}
+              onCaptionsChange={this.toggleCaptions}
+              onPlay={this.handlePlay}
+              onPlayheadDown={this.handlePlayheadDown}
+              onPlayheadUp={this.handlePlayheadUp}
+              onPause={this.handlePause}
+              onFullscreenChange={this.toggleFullscreen}
+              onVolumeChange={this.handleVolumeChange}
+              playing={playing}
+              seek={this.seek}
+              volume={volume}
+            />
+          )}
+        </ColorSchemeProvider>
       </div>
     );
   }