+++ packages/gestalt/src/icons/index.js
@@ -115,6 +115,7 @@ import pin from './pin.svg';
 import pinHide from './pin-hide.svg';
 import pinterest from './pinterest.svg';
 import play from './play.svg';
+import protect from './protect.svg';
 import questionMark from './question-mark.svg';
 import refresh from './refresh.svg';
 import remove from './remove.svg';
@@ -285,6 +286,7 @@ export default {
   'pin-hide': pinHide,
   pinterest,
   play,
+  protect,
   refresh,
   'question-mark': questionMark,
   remove,